tut-gruz-back/pyproject.toml
[project]
name = "tut-gruz-back"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.23.0",
    "fastapi>=0.127.0",
    "pydantic-settings>=2.12.0",
    "uvicorn>=0.40.0",
]
